+++ test/faker/aws_test.exs
@@ -1,6 +1,5 @@
 defmodule Faker.AwsTest do
-    use ExUnit.Case, async: true
-  
-    doctest Faker.Aws.En
-  end
-  
+  use ExUnit.Case, async: true
+
+  doctest Faker.Aws.En
+end